requirements.txt
adjustText==1.3.0
annotated-types==0.7.0
anthropic==0.55.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.8
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==1.26.4
openai==1.91.0
packaging==25.0
pandas==2.3.0
patsy==1.0.1
pillow==11.2.1
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scipy==1.11.4
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
statsmodels==0.14.4
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
